requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
gunicorn
django-cors-headers
